% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{vector of iter}

\item{x}{vector of x}

\item{fun}{function to be applied}

\item{alpha}{alpha value}

\item{cx}{vector for cx}

\item{...}{extra argument}
}
\value{
a plot and list of information about the given sample
}
\description{
Myboot2
}
\examples{
myboot2(x=c(1,1,1,2,3,4,5))
}
