% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{CLT Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{vector of sample size}

\item{iter}{iterator value}

\item{lambda}{lambda value}

\item{...}{extra parameters}
}
\value{
CLT Poisson graph
}
\description{
CLT Poisson
}
\examples{
mycltp(2,10000,4)
}
