% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{a vector for input}

\item{mu}{the mu for the input}

\item{sig}{the sigma for the input}

\item{...}{extra parameters}
}
\value{
a list of mu and sigma
}
\description{
mymlnorm
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
