% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean vector}

\item{sigma}{sd vector}

\item{a}{area vector}
}
\value{
list
}
\description{
myncurve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
