% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothlineplot.R
\name{smoothlineplot}
\alias{smoothlineplot}
\title{Smooth line plot}
\usage{
smoothlineplot(data, x_input, y_input, col_input)
}
\arguments{
\item{data}{data frame to work with}

\item{x_input}{the x value for the data}

\item{y_input}{the y value for the data}

\item{col_input}{what the color should be for the plot}
}
\value{
Plot with smooth lines connecting the data points
}
\description{
Smooth line plot
}
\examples{
smoothlineplot(data=data.frame(Years = 2000:2024),x_input=5,y_input=5,col_input=10)
}
